fix bug of cycle moves overflow
- when passing number of moves to C, we used to pass an integer n, which
  was cycle time times number of atoms (if set in cycle moves).  This
  overflowed.  Fixed by changing it to a double type.

darcs-hash:20080924230522-51e1e-1d0e4307e000c950a89abcbdfd313174d74132d2.gz

+++ ccode/birolimezard.c
@@ -149,7 +149,7 @@ int EddBM_consistencyCheck(struct SimData *SD) {
 }
 
 
-int EddBM_cycle(struct SimData *SD, int n) {
+int EddBM_cycle(struct SimData *SD, double n) {
   if (SD->MLLlen == 0) {
     printf("EddBM_cycle: error, move list length is zero\n");
     exit(12);
